Refactored game engine logic Added Legendary Villains UI changes

+++ LegendaryTestExpansion/BystanderX.cs
@@ -9,6 +9,6 @@ namespace LegendaryTestExpansion
 {
     public class BystanderX : DefaultBystander
     {
-        public override int PrintedVictoryPoints => 2;
+        public override float PrintedVictoryPoints => 2;
     }
 }
